requirements.txt
fastapi[standard]~=0.128.0
pyjwt[crypto]
cryptography
